@@ -1,6 +1,6 @@
 {
   "name": "vue-toast-notification",
-  "version": "0.6.2",
+  "version": "2.0.0",
   "description": "Vue.js toast notification plugin",
   "main": "dist/index.js",
   "types": "./types/toast.d.ts",
@@ -39,61 +39,37 @@
   },
   "dependencies": {},
   "devDependencies": {
-    "@babel/core": "^7.14.3",
-    "@babel/preset-env": "^7.14.4",
-    "@vue/test-utils": "^1.2.0",
-    "babel-core": "7.0.0-bridge.0",
-    "babel-jest": "^27.0.2",
+    "@babel/core": "^7.14.6",
+    "@babel/preset-env": "^7.14.7",
+    "@vue/compiler-sfc": "^3.1.4",
+    "@vue/test-utils": "^2.0.0-rc.10",
+    "babel-jest": "^27.0.6",
     "babel-loader": "^8.2.2",
     "bootstrap": "^4.5.3",
     "clean-webpack-plugin": "^4.0.0-alpha.0",
     "cross-env": "^7.0.3",
     "css-loader": "^5.2.6",
     "file-loader": "^6.2.0",
-    "html-webpack-plugin": "^5.3.1",
-    "jest": "^27.0.4",
-    "mini-css-extract-plugin": "^1.6.0",
-    "mitt": "^2.1.0",
-    "sass": "~1.34.1",
-    "sass-loader": "^12.0.0",
-    "style-loader": "^2.0.0",
+    "html-webpack-plugin": "^5.3.2",
+    "jest": "^27.0.6",
+    "mini-css-extract-plugin": "^2.1.0",
+    "mitt": "^3.0.0",
+    "sass": "~1.35.2",
+    "sass-loader": "^12.1.0",
+    "style-loader": "^3.0.0",
     "svg-url-loader": "^7.1.1",
-    "vue": "^2.6.13",
-    "vue-jest": "^3.0.7",
-    "vue-loader": "^15.9.7",
-    "vue-template-compiler": "^2.6.13",
-    "webpack": "^5.38.1",
-    "webpack-cli": "^4.7.0",
+    "vue": "^3.1",
+    "vue-loader": "^16.3.0",
+    "vue3-jest": "^27.0.0-alpha.1",
+    "webpack": "^5.44.0",
+    "webpack-cli": "^4.7.2",
     "webpack-dev-server": "^4.0.0-beta.3",
     "webpack-remove-empty-scripts": "^0.7.1"
   },
   "peerDependencies": {
-    "vue": "^2.0.0"
+    "vue": "^3.0"
   },
   "engines": {
-    "node": ">=10.13.0"
-  },
-  "jest": {
-    "testEnvironment": "jsdom",
-    "moduleFileExtensions": [
-      "js",
-      "json",
-      "jsx",
-      "node",
-      "vue",
-      "css"
-    ],
-    "setupFiles": [
-      "<rootDir>/__test__/setup.js"
-    ],
-    "transform": {
-      "^.+\\.js$": "babel-jest",
-      ".*\\.(vue)$": "vue-jest"
-    },
-    "moduleNameMapper": {
-      "\\.(css|less|scss)$": "<rootDir>/__test__/__mocks__/styleMock.js"
-    },
-    "collectCoverage": false,
-    "testURL": "http://localhost"
+    "node": ">=12.15.0"
   }
 }
