@@ -1,5 +1,5 @@
 import {mount} from '@vue/test-utils';
-import Component from '../src/js/Component';
+import Component from '../src/js/Component.vue';
 
 describe('Toast component', () => {
   let wrapper;
@@ -8,11 +8,17 @@ describe('Toast component', () => {
     wrapper = mount(Component, {
       propsData: {
         message: 'Test message'
+      },
+      global: {
+        stubs: {
+          transition: false
+        },
       }
     });
   });
 
-  test('renders when active', () => {
+  test('renders when active', async () => {
+    await wrapper.vm.$nextTick();
     expect(wrapper.vm.$el).toMatchSnapshot()
   });
 
