@@ -18,7 +18,7 @@ module.exports = {
       path.resolve(__dirname, 'node_modules'),
     ],
     alias: {
-      'vue$': 'vue/dist/vue.runtime.esm.js'
+      vue: "@vue/runtime-dom"
     },
     extensions: ['.js', '.jsx', '.json', '.vue'],
   },
